fix: remove file representation interface
