+++ libs/vre/resource-editor/representations/src/index.ts
@@ -11,7 +11,7 @@ export * from './lib/text/text.component';
 export * from './lib/video/video.component';
 export * from './lib/video-preview/video-preview.component';
 export * from './lib/representation.service';
-export * from './lib/file-representation';
+export * from './lib/representation-constants';
 export * from './lib/file-representation.type';
 export * from './lib/region';
 export * from './lib/region.service';
