+++ libs/vre/resource-editor/representations/src/lib/representation-constants.ts
@@ -1,0 +1,11 @@
+import { Constants } from '@dasch-swiss/dsp-js';
+
+export class RepresentationConstants {
+  static audio = Constants.AudioFileValue;
+  static document = Constants.DocumentFileValue;
+  static movingImage = Constants.MovingImageFileValue;
+  static stillImage = Constants.StillImageFileValue;
+  static externalStillImage = Constants.StillImageExternalFileValue;
+  static archive = Constants.ArchiveFileValue;
+  static text = Constants.TextFileValue;
+}
