+++ libs/vre/resource-editor/resource-editor/src/lib/resource-representation.component.ts
@@ -1,14 +1,11 @@
 import { Component, Input, OnChanges } from '@angular/core';
-import {
-  FileRepresentation,
-  getFileValue,
-  RepresentationConstants,
-} from '@dasch-swiss/vre/resource-editor/representations';
+import { ReadFileValue } from '@dasch-swiss/dsp-js';
+import { getFileValue, RepresentationConstants } from '@dasch-swiss/vre/resource-editor/representations';
 import { DspResource } from '@dasch-swiss/vre/shared/app-common';
 
 @Component({
   selector: 'app-resource-representation',
-  template: ` <div class="representation-container center" [ngSwitch]="representationToDisplay.fileValue.type">
+  template: ` <div class="representation-container center" [ngSwitch]="fileValue.type">
     <app-still-image
       #stillImage
       class="dsp-representation stillimage"
@@ -25,48 +22,48 @@ import { DspResource } from '@dasch-swiss/vre/shared/app-common';
     <app-document
       #document
       class="dsp-representation document"
-      [class.pdf]="representationToDisplay.fileValue.filename.split('.').pop() === 'pdf'"
+      [class.pdf]="fileValue.filename.split('.').pop() === 'pdf'"
       *ngSwitchCase="representationConstants.document"
-      [src]="representationToDisplay.fileValue"
+      [src]="fileValue"
       [parentResource]="resource.res" />
 
     <app-audio
       #audio
       class="dsp-representation audio"
       *ngSwitchCase="representationConstants.audio"
-      [src]="representationToDisplay.fileValue"
+      [src]="fileValue"
       [parentResource]="resource.res" />
 
     <app-video
       #video
       class="dsp-representation video"
       *ngSwitchCase="representationConstants.movingImage"
-      [src]="representationToDisplay.fileValue"
+      [src]="fileValue"
       [parentResource]="resource.res" />
 
     <app-archive
       #archive
       class="dsp-representation archive"
       *ngSwitchCase="representationConstants.archive"
-      [src]="representationToDisplay.fileValue"
+      [src]="fileValue"
       [parentResource]="resource.res" />
 
     <app-text
       #text
       class="dsp-representation text"
       *ngSwitchCase="representationConstants.text"
-      [src]="representationToDisplay.fileValue"
+      [src]="fileValue"
       [parentResource]="resource.res" />
   </div>`,
 })
 export class ResourceRepresentationComponent implements OnChanges {
   @Input({ required: true }) resource!: DspResource;
 
-  representationToDisplay!: FileRepresentation;
+  fileValue!: ReadFileValue;
   loading = false;
   protected readonly representationConstants = RepresentationConstants;
 
   ngOnChanges() {
-    this.representationToDisplay = new FileRepresentation(getFileValue(this.resource.res)!);
+    this.fileValue = getFileValue(this.resource.res);
   }
 }
